working procedures, loops etc

+++ src/main/ExternalProcedures.java
@@ -2,11 +2,11 @@ package main;
 
 public class ExternalProcedures {
 
-    public static void print(int value){
+    public static void print(long value){
         System.out.println(value);
     }
 
-    public static void print_float(float value){
+    public static void print_float(double value){
         System.out.println(value);
     }
 }
